refactor: refactored entities to models folder and requests and responses bodies to dtos folder
