+++ src/models/users.go
@@ -1,4 +1,4 @@
-package db
+package models
 
 import "gorm.io/gorm"
